ENH: separate configuration options from glbl.py
WIP: working on logic of reloading glbl

fail to reload as after dirs are created, glbl flush default contents

FIX: fix sequence of getting shutter
